+++ routing/http/server/server_test.go
@@ -1,16 +1,22 @@
 package server
 
 import (
+	"bytes"
 	"context"
+	"crypto/rand"
 	"io"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 	"time"
 
+	"github.com/ipfs/boxo/coreiface/path"
+	"github.com/ipfs/boxo/ipns"
+	ipfspath "github.com/ipfs/boxo/path"
 	"github.com/ipfs/boxo/routing/http/types"
 	"github.com/ipfs/boxo/routing/http/types/iter"
 	"github.com/ipfs/go-cid"
+	"github.com/libp2p/go-libp2p/core/crypto"
 	"github.com/libp2p/go-libp2p/core/peer"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -117,6 +123,126 @@ func TestResponse(t *testing.T) {
 	})
 }
 
+func makeName(t *testing.T) (crypto.PrivKey, ipns.Name) {
+	sk, _, err := crypto.GenerateEd25519Key(rand.Reader)
+	require.NoError(t, err)
+
+	pid, err := peer.IDFromPrivateKey(sk)
+	require.NoError(t, err)
+
+	return sk, ipns.NameFromPeer(pid)
+}
+
+func makeIPNSRecord(t *testing.T, cid cid.Cid, sk crypto.PrivKey) (*ipns.Record, []byte) {
+	path := path.IpfsPath(cid)
+	eol := time.Now().Add(time.Hour * 48)
+	ttl := time.Second * 20
+
+	record, err := ipns.NewRecord(sk, ipfspath.FromString(path.String()), 1, eol, ttl)
+	require.NoError(t, err)
+
+	rawRecord, err := ipns.MarshalRecord(record)
+	require.NoError(t, err)
+
+	return record, rawRecord
+}
+
+func TestIPNS(t *testing.T) {
+	cid1, err := cid.Decode("bafkreifjjcie6lypi6ny7amxnfftagclbuxndqonfipmb64f2km2devei4")
+	require.NoError(t, err)
+
+	sk, name1 := makeName(t)
+	record1, rawRecord1 := makeIPNSRecord(t, cid1, sk)
+
+	_, name2 := makeName(t)
+
+	makeRequest := func(t *testing.T, router *mockContentRouter, path string) *http.Response {
+		server := httptest.NewServer(Handler(router))
+		t.Cleanup(server.Close)
+		serverAddr := "http://" + server.Listener.Addr().String()
+		urlStr := serverAddr + path
+		req, err := http.NewRequest(http.MethodGet, urlStr, nil)
+		require.NoError(t, err)
+		req.Header.Set("Accept", mediaTypeIPNSRecord)
+		resp, err := http.DefaultClient.Do(req)
+		require.NoError(t, err)
+		return resp
+	}
+
+	t.Run("GET /routing/v1/ipns/{cid-peer-id} returns 200", func(t *testing.T) {
+		t.Parallel()
+
+		rec, err := ipns.UnmarshalRecord(rawRecord1)
+		require.NoError(t, err)
+
+		router := &mockContentRouter{}
+		router.On("FindIPNSRecord", mock.Anything, name1).Return(rec, nil)
+
+		resp := makeRequest(t, router, "/routing/v1/ipns/"+name1.String())
+		require.Equal(t, 200, resp.StatusCode)
+		require.Equal(t, mediaTypeIPNSRecord, resp.Header.Get("Content-Type"))
+		require.NotEmpty(t, resp.Header.Get("Etag"))
+		require.Equal(t, "max-age=20", resp.Header.Get("Cache-Control"))
+
+		body, err := io.ReadAll(resp.Body)
+		require.NoError(t, err)
+		require.Equal(t, body, rawRecord1)
+	})
+
+	t.Run("GET /routing/v1/ipns/{non-peer-cid} returns 400", func(t *testing.T) {
+		t.Parallel()
+		router := &mockContentRouter{}
+		resp := makeRequest(t, router, "/routing/v1/ipns/"+cid1.String())
+		require.Equal(t, 400, resp.StatusCode)
+	})
+
+	t.Run("GET /routing/v1/ipns/{peer-id} returns 400", func(t *testing.T) {
+		t.Parallel()
+		router := &mockContentRouter{}
+		resp := makeRequest(t, router, "/routing/v1/ipns/"+name1.Peer().String())
+		require.Equal(t, 400, resp.StatusCode)
+	})
+
+	t.Run("PUT /routing/v1/ipns/{cid-peer-id} returns 200", func(t *testing.T) {
+		t.Parallel()
+
+		router := &mockContentRouter{}
+		router.On("ProvideIPNSRecord", mock.Anything, name1, record1).Return(nil)
+
+		server := httptest.NewServer(Handler(router))
+		t.Cleanup(server.Close)
+		serverAddr := "http://" + server.Listener.Addr().String()
+		urlStr := serverAddr + "/routing/v1/ipns/" + name1.String()
+
+		req, err := http.NewRequest(http.MethodPut, urlStr, bytes.NewReader(rawRecord1))
+		require.NoError(t, err)
+		req.Header.Set("Content-Type", mediaTypeIPNSRecord)
+
+		resp, err := http.DefaultClient.Do(req)
+		require.NoError(t, err)
+		require.Equal(t, 200, resp.StatusCode)
+	})
+
+	t.Run("PUT /routing/v1/ipns/{cid-peer-id} returns 400 for wrong record", func(t *testing.T) {
+		t.Parallel()
+
+		router := &mockContentRouter{}
+
+		server := httptest.NewServer(Handler(router))
+		t.Cleanup(server.Close)
+		serverAddr := "http://" + server.Listener.Addr().String()
+		urlStr := serverAddr + "/routing/v1/ipns/" + name2.String()
+
+		req, err := http.NewRequest(http.MethodPut, urlStr, bytes.NewReader(rawRecord1))
+		require.NoError(t, err)
+		req.Header.Set("Content-Type", mediaTypeIPNSRecord)
+
+		resp, err := http.DefaultClient.Do(req)
+		require.NoError(t, err)
+		require.Equal(t, 400, resp.StatusCode)
+	})
+}
+
 type mockContentRouter struct{ mock.Mock }
 
 func (m *mockContentRouter) FindProviders(ctx context.Context, key cid.Cid, limit int) (iter.ResultIter[types.ProviderResponse], error) {
@@ -131,4 +257,14 @@ func (m *mockContentRouter) ProvideBitswap(ctx context.Context, req *BitswapWrit
 func (m *mockContentRouter) Provide(ctx context.Context, req *WriteProvideRequest) (types.ProviderResponse, error) {
 	args := m.Called(ctx, req)
 	return args.Get(0).(types.ProviderResponse), args.Error(1)
+}
+
+func (m *mockContentRouter) FindIPNSRecord(ctx context.Context, name ipns.Name) (*ipns.Record, error) {
+	args := m.Called(ctx, name)
+	return args.Get(0).(*ipns.Record), args.Error(1)
+}
+
+func (m *mockContentRouter) ProvideIPNSRecord(ctx context.Context, name ipns.Name, record *ipns.Record) error {
+	args := m.Called(ctx, name, record)
+	return args.Error(0)
 }